+++ CHANGES.rst
@@ -1,4 +1,4 @@
 3.0.0 (2025-05-13)
 ----------------
 
-- Third and improved version of FLEET.
+- Third and improved version of fleet.